+++ src/components/catalog-item/catalog-item.test.jsx
